Initial commit almost done will finish refactoring later
